requirements.txt
PyYAML==6.0.3
jsonschema==4.23.0
requests==2.32.5
python-dateutil==2.9.0.post0
rich==14.2.0
tenacity==9.1.2
rapidfuzz==3.13.0
watchdog==5.0.1
kubernetes==29.0.0
